working on porting attached images from readme into modal
