+++ src/data/BiopathDetails/biopath-architecture.png
