+++ src/sampleInfoBlob.json
@@ -67,7 +67,7 @@
                     {"name": "REST API"}
                 ],
                 "importanceRanking": 1,
-                "customReadme": ""
+                "customReadmePath": "BiopathDetails"
             },
             {
                 "name": "Reddit r/AITA NLP Sentiment Classifier",
